+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Notify controller about replication state changes in DRBD
 - Notify controller about done percentage changes in DRBD
 
+### Changed
+
+- BalanceResource: Ignore resources which are in a non-valid replication state
+
 ## [1.30.4] - 2025-02-03
 
 ### Changed